# apps/backend/docs/external/player-api/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/players/wallet": {
      "get": {
        "operationId": "WalletController_getBalance_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get player wallet balance",
        "tags": [
          "Wallet"
        ]
      }
    },
    "/api/v1/players/wallet/transactions": {
      "get": {
        "operationId": "WalletController_getTransactions_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transaction history",
        "tags": [
          "Wallet"
        ]
      }
    },
...
  },
  "info": {
    "title": "BTV Player API",
    "description": "--",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:3000"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}